+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Changed web-frontend `/api/docs` route into `/api-docs`.
 * Bumped the dependencies.
 * The internal setting allowing Baserow to run with the user tables in a separate 
   database has been removed entirely to prevent data integrity issues.